Breaking change to pass a PusherEvent object to onEvent
Introduces a PusherEvent class which is now passed as a single argument to onEvent. onEvent used to take 3 arguments (channel name, event name, data), so this is a breaking change.

The PusherEvent exposes a method getUserId which exposes user_ids which are added to client events on presence channels. This is a new feature, and the motivation for this breaking change.