Downgrade gradle to 5.0
The combination of this gradle version, the signing plugin and the
legacy maven plugin seem to cause the error:

Execution failed for task ':signArchives'.
> Duplicate key pusher-websocket-java:jar.asc:asc: (attempted merging values Signature pusher-websocket-java:jar.asc:asc: and Signature pusher-websocket-java:jar.asc:asc:)

There is not much information about this online, except for

https://github.com/gradle/gradle/issues/8213 and
https://discuss.gradle.org/t/apply-plugin-signing-broke-in-gradle-5-1-1-and-prints-out-non-existent-task/31418

Switching to the recommended maven plugin `maven-publish` might
solve this problem, but that will be more work, so downgrading
gradle is a better option in the short term.