Forward channel_data with subscription on private channels
If the auth endpoint returns channel_data, then this should be
forwarded to the auth endpoint even if it is not used by private
channels currently.

This is useful because it allows the same auth endpoint to be
used for private and presence channels.

It also avoids confusing auth signature errors that occur when the
channel_data is not sent (but the auth signature was computed from
it).

Fixes #144 #144 #144.

In order to support the getMe() method on PresenceChannel, we need
to store the channel_data in the PrivateChannelImpl class. This
is decoded in the PresenceChannelImpl class since presence channels
require the channel_data is JSON. It is decoded in toSubscribeMessage()
rather than getMe() to preserve the original semantics (e.g.
exceptions cannot be thrown from getMe().