Do not use the disconnect method to tear down on pong timeout
The `disconnect` method is public, and tests the current state before
taking action, scheduling this work to happen on the event thread.

However, when a pong times out, the `onClose` is executed immediately,
transitioning in to the reconnecting state before the scheduled
disconnect happens, meaning disconnect fails to take any action and
potentially resulting in the old connection not being torn down.